fix: to_cnf and to_dnf functions
